[OPE-47] Colocate concrete RHIME recipes
