+++ openghg_inversions/config/templates/rhime_template.ini
@@ -69,8 +69,6 @@ chains = 4
 averaging_error = True
 min_error = 0.0
 fix_basis_outer_regions = False
-; Use "compiled" to opt into the experimental semantic-plan compiler.
-builder_strategy = "concrete"
 ; Select "dense", "low_rank", or "diagonal" to override prepared-input auto-detection.
 aggregation_error_mode = "auto"
 ; "lazy" records and applies zero-fill lazily; "count" computes exact non-finite flux counts once.
