+++ openghg_inversions/hbmcmc/inversion_pymc.py
@@ -24,7 +24,7 @@ from openghg_inversions import utils  # noqa: E402
 from openghg_inversions._sampling import _reset_retained_draws  # noqa: E402
 from openghg_inversions.hbmcmc.hbmcmc_output import define_output_filename  # noqa: E402
 from openghg_inversions.config.version import code_version  # noqa: E402
-from openghg_inversions.models import build_rhime_model  # noqa: E402
+from openghg_inversions.rhime.standard import build_rhime_model  # noqa: E402
 from openghg_inversions.models.components import resolve_model_variable  # noqa: E402
 from openghg_inversions.models.coords import get_coord_registry, restore_inferencedata_coords  # noqa: E402
 from openghg_inversions.models.priors import PriorArgs  # noqa: E402
