+++ openghg_inversions/models/__init__.py
@@ -47,13 +47,8 @@ from openghg_inversions.models.rhime import (
     DEFAULT_OFFSET_PRIOR,
     DEFAULT_SIGMA_PRIOR,
     DEFAULT_X_PRIOR,
-    RhimeBuilderStrategy,
     RhimeModelSpec,
     SectorSpec,
-    build_rhime_model,
-    build_rhime_model_from_spec,
-    build_rhime_multisector_model,
-    build_rhime_multisector_model_from_spec,
     get_rhime_likelihood_result,
     safe_pymc_name,
 )
@@ -78,7 +73,6 @@ __all__ = [
     "DEFAULT_SIGMA_PRIOR",
     "DEFAULT_X_PRIOR",
     "AggregationErrorMode",
-    "RhimeBuilderStrategy",
     "RhimeLikelihoodBuilder",
     "RhimeLikelihoodContext",
     "RhimeLikelihoodResult",
@@ -105,10 +99,6 @@ __all__ = [
     "build_absolute_sigma_gaussian_likelihood",
     "build_gaussian_rhime_likelihood",
     "build_rhime_observation_state",
-    "build_rhime_model",
-    "build_rhime_model_from_spec",
-    "build_rhime_multisector_model",
-    "build_rhime_multisector_model_from_spec",
     "get_rhime_likelihood_result",
     "safe_pymc_name",
     "active_prior_args",
