+++ openghg_inversions/rhime/multisector.py
@@ -2,21 +2,38 @@
 
 from __future__ import annotations
 
+from collections.abc import Mapping, Sequence
 from pathlib import Path
-from typing import Any
+from typing import Any, cast
 
 import arviz as az
+import pymc as pm
+import pytensor.tensor as pt
 import xarray as xr
 
 from openghg_inversions._timing import log_timing, timer_seconds, timer_start
 from openghg_inversions.basis.basis_functions import BasisFunctions
 from openghg_inversions.inversion_data import RhimeMergedData, RhimePreparedInputs
-from openghg_inversions.models import (
-    RhimeLikelihoodBuilder,
-    RhimeModelSpec,
-    build_rhime_multisector_model_from_spec,
+from openghg_inversions.models._rhime_flux import (
+    _resolve_multisector_components,
+    _resolve_sector_bindings,
+    _select_sector_design,
 )
-from openghg_inversions.models._rhime_flux import _select_sector_design
+from openghg_inversions.models.components import add_state_linear_component
+from openghg_inversions.models.coords import CoordRegistry, attach_coord_registry
+from openghg_inversions.models.rhime import (
+    DEFAULT_BC_PRIOR,
+    DEFAULT_OFFSET_PRIOR,
+    DEFAULT_SIGMA_PRIOR,
+    DEFAULT_X_PRIOR,
+    RhimeModelSpec,
+    _add_rhime_observation_components,
+    _LIKELIHOOD_RESULT_ATTR,
+)
+from openghg_inversions.models.rhime_likelihood import RhimeLikelihoodBuilder
+from openghg_inversions.models.state_activity import StateActivity
+from openghg_inversions.observation_error import AggregationErrorMode
+from openghg_inversions.sigma import SigmaAlignment
 
 from ._model_building import (
     builtin_model_build_result,
@@ -43,6 +60,142 @@ from .preparation import (
 )
 from .sampling import RhimeSampler, sample_rhime_model
 from .specs import RhimeRunSpec
+
+
+def build_rhime_multisector_model(
+    inv_inputs: xr.Dataset,
+    *,
+    sigma_alignment: SigmaAlignment,
+    sectors: Sequence[str] | None = None,
+    sector_sources: Mapping[str, str] | None = None,
+    sector_variable_suffixes: Mapping[str, str] | None = None,
+    sector_priors: Mapping[str, dict] | None = None,
+    x_prior: dict | None = None,
+    bc_prior: dict | None = None,
+    sigma_prior: dict | None = None,
+    offset_prior: dict | None = None,
+    add_offset: bool = False,
+    use_bc: bool = True,
+    pollution_events_from_obs: bool = False,
+    no_model_error: bool = False,
+    aggregation_error_mode: AggregationErrorMode = "auto",
+    offset_args: dict | None = None,
+    power: dict | float = 1.99,
+    state_activity: StateActivity | None = None,
+    sector_state_activities: Mapping[str, StateActivity] | None = None,
+    bc_state_activity: StateActivity | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> pm.Model:
+    """Build the concrete shared-basis multi-sector RHIME model.
+
+    Each sector receives its own state vector ``x_<sector>`` and forward-model
+    contribution ``mu_<sector>``. The total ``mu`` is the sum of sector
+    contributions and is passed to the standard RHIME likelihood.
+    """
+    sector_bindings = _resolve_sector_bindings(
+        inv_inputs,
+        sectors,
+        sector_sources=sector_sources,
+        sector_variable_suffixes=sector_variable_suffixes,
+    )
+    sector_components = _resolve_multisector_components(
+        inv_inputs,
+        sector_bindings,
+        sector_priors=sector_priors,
+        x_prior=x_prior,
+        default_x_prior=DEFAULT_X_PRIOR,
+        state_activity=state_activity,
+        sector_state_activities=sector_state_activities,
+    )
+    bc_prior = dict(DEFAULT_BC_PRIOR if bc_prior is None else bc_prior)
+    sigma_prior = dict(DEFAULT_SIGMA_PRIOR if sigma_prior is None else sigma_prior)
+    offset_prior = dict(DEFAULT_OFFSET_PRIOR if offset_prior is None else offset_prior)
+    with pm.Model() as model:
+        attach_coord_registry(model, CoordRegistry())
+        sector_outputs = []
+        for component in sector_components:
+            linear_component = add_state_linear_component(
+                component.design,
+                data_name=f"hx_{component.variable_suffix}",
+                prior_args=dict(component.prior_args),
+                var_name=f"x_{component.variable_suffix}",
+                output_name=f"mu_{component.variable_suffix}",
+                output_dim="nmeasure",
+                compute_deterministic=True,
+                state_activity=component.state_activity,
+            )
+            sector_outputs.append(linear_component.output)
+
+        total_mu = pm.Deterministic(
+            "mu",
+            cast(Any, pt.stack(sector_outputs, axis=0)).sum(axis=0),
+            dims="nmeasure",
+        )
+        likelihood_result = _add_rhime_observation_components(
+            inv_inputs,
+            mu=total_mu,
+            sigma_alignment=sigma_alignment,
+            bc_prior=bc_prior,
+            sigma_prior=sigma_prior,
+            offset_prior=offset_prior,
+            add_offset=add_offset,
+            use_bc=use_bc,
+            bc_state_activity=bc_state_activity,
+            pollution_events_from_obs=pollution_events_from_obs,
+            no_model_error=no_model_error,
+            aggregation_error_mode=aggregation_error_mode,
+            offset_args=offset_args,
+            power=power,
+            likelihood_builder=likelihood_builder,
+        )
+        setattr(model, _LIKELIHOOD_RESULT_ATTR, likelihood_result)
+
+    return model
+
+
+def build_rhime_multisector_model_from_spec(
+    inv_inputs: xr.Dataset,
+    model_spec: RhimeModelSpec,
+    *,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> pm.Model:
+    """Build the concrete multi-sector RHIME model from a model spec."""
+    sigma_alignment = SigmaAlignment.from_frequency(
+        inv_inputs["site_indicator"],
+        frequency=model_spec.sigma_freq,
+        per_site=model_spec.sigma_per_site,
+        anchor_time=model_spec.sigma_freq_anchor,
+    )
+    sector_state_activities = dict(model_spec.sector_state_activities or {})
+    sector_state_activities.update(
+        {
+            sector.name: sector.state_activity
+            for sector in model_spec.sectors
+            if sector.state_activity is not None
+        }
+    )
+    return build_rhime_multisector_model(
+        inv_inputs,
+        sigma_alignment=sigma_alignment,
+        sectors=[sector.name for sector in model_spec.sectors],
+        sector_sources={sector.name: sector.flux_source for sector in model_spec.sectors},
+        sector_variable_suffixes={sector.name: sector.variable_suffix for sector in model_spec.sectors},
+        sector_priors={sector.name: dict(sector.x_prior) for sector in model_spec.sectors},
+        bc_prior=model_spec.bc_prior,
+        bc_state_activity=model_spec.bc_state_activity,
+        sigma_prior=model_spec.sigma_prior,
+        offset_prior=model_spec.offset_prior,
+        add_offset=model_spec.add_offset,
+        use_bc=model_spec.use_bc,
+        pollution_events_from_obs=model_spec.pollution_events_from_obs,
+        no_model_error=model_spec.no_model_error,
+        aggregation_error_mode=model_spec.aggregation_error_mode,
+        offset_args=model_spec.offset_args,
+        power=model_spec.power,
+        state_activity=model_spec.state_activity,
+        sector_state_activities=sector_state_activities or None,
+        likelihood_builder=likelihood_builder,
+    )
 
 
 def _validate_multisector_basis_layout(
