+++ openghg_inversions/rhime/standard.py
@@ -11,11 +11,23 @@ from pathlib import Path
 from typing import Any
 
 import arviz as az
+import pymc as pm
 import xarray as xr
 
 from openghg_inversions._timing import log_timing, timer_seconds, timer_start
 from openghg_inversions.inversion_data import RhimeMergedData, RhimePreparedInputs
-from openghg_inversions.models import RhimeLikelihoodBuilder, build_rhime_model_from_spec
+from openghg_inversions.models.components import add_state_linear_component
+from openghg_inversions.models.coords import CoordRegistry, attach_coord_registry
+from openghg_inversions.models.rhime import (
+    _LIKELIHOOD_RESULT_ATTR,
+    _add_rhime_observation_components,
+    _prepare_builder_priors,
+    RhimeModelSpec,
+)
+from openghg_inversions.models.rhime_likelihood import RhimeLikelihoodBuilder
+from openghg_inversions.models.state_activity import StateActivity
+from openghg_inversions.observation_error import AggregationErrorMode
+from openghg_inversions.sigma import SigmaAlignment
 
 from ._model_building import (
     builtin_model_build_result,
@@ -42,6 +54,114 @@ from .preparation import (
 )
 from .sampling import RhimeSampler, sample_rhime_model
 from .specs import RhimeRunSpec
+
+
+def build_rhime_model(
+    inv_inputs: xr.Dataset,
+    *,
+    sigma_alignment: SigmaAlignment,
+    x_prior: dict | None = None,
+    bc_prior: dict | None = None,
+    sigma_prior: dict | None = None,
+    offset_prior: dict | None = None,
+    add_offset: bool = False,
+    use_bc: bool = True,
+    pollution_events_from_obs: bool = False,
+    no_model_error: bool = False,
+    aggregation_error_mode: AggregationErrorMode = "auto",
+    offset_args: dict | None = None,
+    power: dict | float = 1.99,
+    state_activity: StateActivity | None = None,
+    bc_state_activity: StateActivity | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> pm.Model:
+    """Build the concrete standard single-sector RHIME model.
+
+    The recipe constructs the flux contribution first, then adds the shared
+    baseline, offset, model-data mismatch, aggregation-error, and likelihood
+    components in scientific order.
+    """
+    x_prior, bc_prior, sigma_prior, offset_prior = _prepare_builder_priors(
+        x_prior=x_prior,
+        bc_prior=bc_prior,
+        sigma_prior=sigma_prior,
+        offset_prior=offset_prior,
+    )
+
+    with pm.Model() as model:
+        attach_coord_registry(model, CoordRegistry())
+        flux_component = add_state_linear_component(
+            inv_inputs["H"],
+            data_name="hx",
+            prior_args=x_prior,
+            var_name="x",
+            output_name="mu",
+            output_dim="nmeasure",
+            compute_deterministic=True,
+            state_activity=state_activity,
+        )
+        likelihood_result = _add_rhime_observation_components(
+            inv_inputs,
+            mu=flux_component.output,
+            sigma_alignment=sigma_alignment,
+            bc_prior=bc_prior,
+            sigma_prior=sigma_prior,
+            offset_prior=offset_prior,
+            add_offset=add_offset,
+            use_bc=use_bc,
+            bc_state_activity=bc_state_activity,
+            pollution_events_from_obs=pollution_events_from_obs,
+            no_model_error=no_model_error,
+            aggregation_error_mode=aggregation_error_mode,
+            offset_args=offset_args,
+            power=power,
+            likelihood_builder=likelihood_builder,
+        )
+        setattr(model, _LIKELIHOOD_RESULT_ATTR, likelihood_result)
+
+    return model
+
+
+def build_rhime_model_from_spec(
+    inv_inputs: xr.Dataset,
+    model_spec: RhimeModelSpec,
+    *,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> pm.Model:
+    """Build the standard model from a normalized RHIME model specification."""
+    if len(model_spec.sectors) != 1:
+        raise ValueError("Standard RHIME model specs must include exactly one sector.")
+
+    sector = model_spec.sectors[0]
+    sigma_alignment = SigmaAlignment.from_frequency(
+        inv_inputs["site_indicator"],
+        frequency=model_spec.sigma_freq,
+        per_site=model_spec.sigma_per_site,
+        anchor_time=model_spec.sigma_freq_anchor,
+    )
+    state_activity = model_spec.state_activity
+    if model_spec.sector_state_activities is not None:
+        state_activity = model_spec.sector_state_activities.get(sector.name, state_activity)
+    if sector.state_activity is not None:
+        state_activity = sector.state_activity
+    return build_rhime_model(
+        inv_inputs,
+        sigma_alignment=sigma_alignment,
+        x_prior=dict(sector.x_prior),
+        state_activity=state_activity,
+        bc_prior=model_spec.bc_prior,
+        bc_state_activity=model_spec.bc_state_activity,
+        sigma_prior=model_spec.sigma_prior,
+        offset_prior=model_spec.offset_prior,
+        add_offset=model_spec.add_offset,
+        use_bc=model_spec.use_bc,
+        pollution_events_from_obs=model_spec.pollution_events_from_obs,
+        no_model_error=model_spec.no_model_error,
+        aggregation_error_mode=model_spec.aggregation_error_mode,
+        offset_args=model_spec.offset_args,
+        power=model_spec.power,
+        likelihood_builder=likelihood_builder,
+    )
 
 
 def build_standard_rhime_model(
